requirements.txt
requests>=2.25.1
Pillow>=8.0.0
python-whois>=0.8.0
